requirements.txt
fastapi
uvicorn
python-pptx
